+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_responses.hpp
@@ -51,8 +51,11 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
   using DataLakeSignedIdentifier = Blobs::Models::BlobSignedIdentifier;
   using ListDataLakeFileSystemsIncludeFlags = Blobs::Models::ListBlobContainersIncludeFlags;
 
-  struct DataLakeFileSystemAccessPolicy
+  struct GetDataLakeFileSystemAccessPolicyResult
   {
+    std::string RequestId;
+    Azure::ETag ETag;
+    Azure::DateTime LastModified;
     PublicAccessType AccessType = PublicAccessType::None;
     std::vector<DataLakeSignedIdentifier> SignedIdentifiers;
   }; // struct DataLakeFileSystemAccessPolciy