+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added support for telemetry options.
 - Added `Azure::Storage::Blobs::PackageVersion`.
+- Added `ShareFileClient::GetRangeListDiff`.
 
 ### Breaking Changes
 
@@ -17,8 +18,9 @@
 - Changed return type of `BlobServiceClient::GetAccountInfo` to `AccountInfo`.
 - Changed return type of `BlobServiceClient::GetStatistics` to `ServiceStatistics`.
 - Changed return type of `BlobContainerClient::GetProperties` to `BlobContainerProperties`.
-- Changed return type of `BlobContainerClient::GetAccessPolicy` to `BlobContainerAccessPolicy`.
 - Changed return type of `BlobClient::GetProperties` to `BlobProperties`.
+- Changed return type of `BlobClinet::GetTags` to `std::map<std::string, std::string>`.
+- Removed `PreviousShareSnapshot` from `GetShareFileRangeListOptions`, use `ShareFileClient::GetRangeListDiff` instead.
 
 ### Other Changes and Improvements